Add files via upload
I utilized this particular dataset, which can be found at https://www.kaggle.com/datasets/meirnizri/covid19-dataset. This dataset is comprehensive and encompasses a variety of variables that are crucial for my analysis. It provides detailed information that enhances the scope of my research, allowing for a more in-depth examination of the subject matter.